+++ src/components/swapv2/SwapModalFooter.tsx
@@ -122,7 +122,7 @@ export default function SwapModalFooter({
             </RowFixed>
             <TYPE.black color={theme.text} fontSize={12}>
               {formattedNum(
-                ((parseFloat(trade.amountInUsd) * parseInt(feeConfig.feeAmount)) / 10000)?.toString(),
+                ((parseFloat(trade.amountInUsd) * parseFloat(feeConfig.feeAmount)) / 10000)?.toString(),
                 true
               )}
             </TYPE.black>
